requirements.txt
# Core LangChain
langchain-core==0.3.52
langchain-community==0.3.0
langchain==0.3.0
langchain-google-genai # Supports Gemini 2.0 Flash

# Google Gemini (Latest)
google-generativeai  # Required for Gemini 2.0 Flash

# Streamlit & Utilities
streamlit==1.33.0
python-dotenv==1.0.1

# Vector DB & Embeddings
faiss-cpu==1.8.0
sentence-transformers==2.7.0

# Document Processing
pypdf==4.1.0
unstructured==0.13.4
